Port "Fix http://bugs.mysql.com/62534 by making the my.cnf innodb_max_dirty_pages_pct variable a double" to 5.6
Summary:
Original (https://phabricator.fb.com/D333501): This makes innodb_max_dirty_pages_pct a double with min,default,max values
0.001, 75, 99.999.
"
Added more to the BUFFER POOL AND MEMORY section of SHOW INNODB STATUS
Percent pages dirty: 8.939 --> this is all n_dirty_pages / used_pages
+Percent all pages dirty: 0.195 --> this is n_dirty_pages / all-pages
+Max dirty pages percent: 75.000 --> this is innodb_max_dirty_pages_pct
"
Also made all of the buf use 3 digits of precision instead of 2 (%.2f -> %.3f)

Extra: Also made innodb_max_dirty_pages_pct_lwm a double sysvar as these variables are inter-dependent.

Test Plan: Jenkins

Reviewers: jtolmer, nizamordulu

Reviewed By: jtolmer

CC: db-eng@lists, steaphan, mcallaghan

Differential Revision: https://phabricator.fb.com/D751157

Task ID: 1957313

+++ storage/innobase/srv/srv0srv.cc
@@ -286,8 +286,8 @@ in the buffer pool to all database pages in the buffer pool smaller than
 the following number. But it is not guaranteed that the value stays below
 that during a time of heavy update/insert activity. */
 
-UNIV_INTERN ulong	srv_max_buf_pool_modified_pct	= 75;
-UNIV_INTERN ulong	srv_max_dirty_pages_pct_lwm	= 50;
+UNIV_INTERN double	srv_max_buf_pool_modified_pct	= 75.0;
+UNIV_INTERN double	srv_max_dirty_pages_pct_lwm	= 50.0;
 
 /* This is the percentage of log capacity at which adaptive flushing,
 if enabled, will kick in. */
